Refs 13570. Moved rearrange to implementation folder.
