+++ src/cpp/fastdds/topic/DDSSQLFilter/DDSFilterExpressionParser.hpp
@@ -36,50 +36,7 @@ namespace parser {
 
 using namespace tao::TAO_PEGTL_NAMESPACE;
 
-// after a node is stored successfully, you can add an optional transformer like this:
-struct rearrange
-    : parse_tree::apply< rearrange >  // allows bulk selection, see selector<...>
-{
-    // recursively rearrange nodes. the basic principle is:
-    //
-    // from:          PROD/EXPR
-    //                /   |   \          (LHS... may be one or more children, followed by OP,)
-    //             LHS... OP   RHS       (which is one operator, and RHS, which is a single child)
-    //
-    // to:               OP
-    //                  /  \             (OP now has two children, the original PROD/EXPR and RHS)
-    //         PROD/EXPR    RHS          (Note that PROD/EXPR has two fewer children now)
-    //             |
-    //            LHS...
-    //
-    // if only one child is left for LHS..., replace the PROD/EXPR with the child directly.
-    // otherwise, perform the above transformation, then apply it recursively until LHS...
-    // becomes a single child, which then replaces the parent node and the recursion ends.
-    template< typename ... States >
-    static void transform(
-            std::unique_ptr< parse_tree::node >& n,
-            States&&... st)
-    {
-        if (n->children.size() == 1)
-        {
-            n = std::move(n->children.back());
-        }
-        else
-        {
-            n->remove_content();
-            auto& c = n->children;
-            auto r = std::move(c.back());
-            c.pop_back();
-            auto o = std::move(c.back());
-            c.pop_back();
-            o->children.emplace_back(std::move(n));
-            o->children.emplace_back(std::move(r));
-            n = std::move(o);
-            transform(n->children.front(), st ...);
-        }
-    }
-
-};
+#include "DDSFilterExpressionParserImpl/rearrange.hpp"
 
 // select which rules in the grammar will produce parse tree nodes:
 template< typename Rule >
